refactor(login): split LoginPreOrchestrator (P12d)
Phase 12d file 6/8 (Login + Form HIGH cluster decoupling).

Why
- LoginPreOrchestrator.ts mixed three orthogonal concerns:
  early gates, readiness/preAction callbacks + neterror probe,
  and post-preamble DOM-ready + field discovery + commit.
- Splitting per the Phase 12d cluster pattern proven by PR #339
  + files 1-5 of this branch.

What
- Extracted PreOrchestrator/PreOrchestratorTypes.ts: failure
  messages (LOGIN_PRE_NO_BROWSER, LOGIN_PRE_NO_MEDIATOR), the
  LOGIN_PRE_FRAME_PRELUDE spec, plus DiscoverFormPreamble +
  IDiscoverFormResources types.
- Extracted PreOrchestrator/PreOrchestratorReadiness.ts:
  probeNeterrorAndFail + runDiscoverFormPreamble (with private
  failCheckReadiness/performCheckReadiness/runCheckReadiness/
  performPreAction/runPreAction helpers).
- Extracted PreOrchestrator/PreOrchestratorDiscovery.ts:
  runPostPreamble (with private buildLoginState/waitFormDomReady/
  buildDiscoverFieldsArgs/runFieldDiscovery/commitDiscoverForm).
- Rewrote LoginPreOrchestrator.ts as a 60-LoC façade that owns
  only runDiscoverFormFlow + executeDiscoverForm (default +
  named export preserved byte-identically).

Guideline compliance
- before-commit-guidlines.md: all 23 husky gates expected GREEN
  (lint, lint:architecture, lint:canaries, tsc, eslint clean,
  147/147 targeted tests, docs-coverage allowlisted).
- eslint-rules-guidlines.md S3: zero eslint-disable headers
  added; the prefer-default-export rule satisfied by re-exporting
  the sibling types DiscoverFormPreamble (Readiness) and
  IDiscoverFormResources (Discovery).
- eslint-rules-guidlines.md S5: no eslint.config.mjs changes;
  refactor commit is separable from any guardrail commit.
- comments-in-code-guidlines.md / j-doc-guidlines.md: BLUF
  JSDoc on every exported + internal helper; no commented-out
  code.
- general-rules-guidlines.md P7 Mediator: executeDiscoverForm
  remains the sole LoginPhase entry point through the mediator
  facade.
- design-patterns-guidlines.md Result Pattern: every async
  boundary returns Procedure<T>; no exception-based control
  flow introduced.

Tracker(s)
- Phase 12d
- PVTI_lAHOAPtF084BZbzwzgvauEI
